Switched to VS2017. Used new csproj format. Splitted JSonSerialization in serveral files. Moves files in subfolders
