+++ MegaApiClient.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.12
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{F2689498-3543-43EA-8961-66385687E16F}"
 	ProjectSection(SolutionItems) = preProject
@@ -25,6 +25,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MegaApiClient.Tests", "Mega
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MegaApiClient-4.5", "MegaApiClient-4.5\MegaApiClient-4.5.csproj", "{19D28272-9DDB-4868-93E7-D98DCFAE08E9}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MegaApiClient", "MegaApiClient\MegaApiClient.csproj", "{F9A241E2-4E26-493D-B07E-48AB86066C05}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,6 +49,10 @@ Global
 		{19D28272-9DDB-4868-93E7-D98DCFAE08E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{19D28272-9DDB-4868-93E7-D98DCFAE08E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{19D28272-9DDB-4868-93E7-D98DCFAE08E9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F9A241E2-4E26-493D-B07E-48AB86066C05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F9A241E2-4E26-493D-B07E-48AB86066C05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F9A241E2-4E26-493D-B07E-48AB86066C05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F9A241E2-4E26-493D-B07E-48AB86066C05}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
