+++ MegaApiClient/Interface/IMegaApiClient.cs
