+++ MegaApiClient/Interface/IMegaApiClientAsync.cs
