+++ MegaApiClient/Interface/IWebClient.cs
