+++ MegaApiClient/Serialization/Logout.cs
@@ -1,0 +1,10 @@
+﻿namespace CG.Web.MegaApiClient.Serialization
+{
+  internal class LogoutRequest : RequestBase
+  {
+    public LogoutRequest()
+      : base("sml")
+    {
+    }
+  }
+}
