+++ MegaApiClient/Stream/CancellableStream.cs
@@ -1,9 +1,9 @@
-using System;
-using System.IO;
-using System.Threading;
-
 namespace CG.Web.MegaApiClient
 {
+  using System;
+  using System.IO;
+  using System.Threading;
+
   public class CancellableStream : Stream
   {
     private Stream stream;
