+++ MegaApiClient/Stream/MegaAesCtrStream.cs
