+++ MegaApiClient/Stream/ProgressionStream.cs
